move hairy filtering out of index.js
